Fix annoying {%--%} %} autoclose issue
VS Code doesn't help much when it comes to recognizing tags that were
already closed. And since folks might want to add/remove the minus sign
on only one side of the tag and there's no clean way of having it be
done on both at once, I'm removing this feature.

Should be less annoying.

Fixes #48