Change the rule for attributes breaking on new lines
I put in > 2 initially but I don't think that's what we want.

Stuff like <link href=... preload=... foo=...> should stay on a line if
that's what we want.

Instead, I changed the rule to "if there was a line break in the input,
break it on the output". This would keep styling as it was and only
break if the line is too long or it already was broken up.