Make isTrailingWhitespaceSensitive & isLeadingWhitespaceSensitive match
i.e. a node is trailing whitespace sensitive iff this one is and the
next one is leading whitespace sensitive.

e.g. a <span> could be trailing WS sensitive a <div> could be leading
insensitive. In the <span></span><div></div> case, the span is thus
trailing insensitive because of the div.