Improve liquid snapshot tests
We'll use lines that start with It|If|When|// as the test descriptions.
This way a file can break for multiple reasons and they will be reported
separately.

Will make it so you're not working on a file by file basis but on a use
case basis.

Temporarily increases the number of failed runs but it's actually
accurate now.