Properly parse LiquidDrop strings
This kicks off the "proper" parsing of Liquid.

Herein, we start extending the ohm grammar to support Liquid constructs
inside the markup of `LiquidDrop`s.

It goes:
```
LiquidDrop {
  markup: LiquidVariable {
    expression: LiquidString
    filters: LiquidFilter[]
  }
}
```

And if we can't parse the string, we fallback to the old implementation
that didn't parse anything.

This means `{{ 'string' }}` will get replaced to `{{ "string" }}` unless
`prettierOptions.singleQuote`