Fixup parsing of `{% else-%}`
This is a weird parsing ambiguity problem.

Before, our rule was the following:

```
identifierCharacter = alnum | "_" | "-"
"{%" (tagName ~identifierCharacter) space* tagMarkup "-"? "%}"
```

Our problem was that `-` is both an identifier character and the character we use for stripping whitespace from the end of tags and variable output (`-%}` and `-}}`).

So `{% else-%}` threw an error because `tagName` _was_ followed by `-`, which is an identifier character.

The solution is to change that rule to the following:
- Followed by an "end of tag" `-}}`, `-%}`, OR
- Not followed by an identifierCharacter

Fixes #126