Add support for fixes and suggestions (#80)
* Add API design for fix & suggest

Heavily inspired by eslint, but also from codemirror and the LSP.

When you write checks and use `context.report`, you can also provide
`fix` and/or `suggest` parameters.

The `fix` property is used when you can automatically fix an problem.
That means that the fix is safe in all cases. It would be suitable for a
`--autofix` flag or for a CodeEditor option that automatically fixes
problems.

The `suggest` property is an array of `Suggestion`. Each `Suggestion`
represent one way of fixing the problem. They are to be used when there
are more than one way to fix the problem, or when an otherwise automatic
fix would be unsafe.

A `Fix` is either a `FixDescription`, or an array of `Fix`. It can thus
be arbitrarily nested and concatenated.

A `FixApplicator` is a function that takes the untouched source of a
file and asynchronously applies the collection of `FixDescriptions` that
it is passed.

In Node.js, for the `--autofix` flag, the `FixApplicator` might apply all the
`FixDescription`s on the source to obtain the final result, and then
persist that to disk with `fs.saveFile`.

In the Language Server, we might have a `FixApplicator` that collects
all the FixDescriptions and turns them into a [`TextEdit` array][1] to
be sent to the LanguageClient in a ApplyWorkspaceEdit request.

[1]: https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textEditArray

* Add ParserBlockingScript suggest example

* Redesign Corrector into a stateful and per SourceCodeType concept

While doing the litmus test for the MatchingTranslations check, I
noticed that the string API for correcting JSON blobs was REALLY REALLY
REALLY awkward.

At the same time, I learned about JSON Patch and how you can describe a
way for a JSON object to mutate with data... Which is pretty much a
perfect use case for us. In the end, I opted out of the spec because it
was too strict for our use case, but the approach is very similar.

In this commit:
- corrector is now an argument of type inferred from the context
  - JSONCorrector for JSONCheckDefinition contexts
  - StringCorrector for LiquidHtmlCheckDefinition contexts
- `Problem` is now `Problem<S>`
- Offense is now `Offense<S>` and has a `type` property
- StaticContextProperties is now `StaticContextProperties<S>`
- Add `expect(offense).to.offerFix()` custom assertion
- Add `expect(offense).to.suggest()` custom assertion

There was a lot of work around making that appear automagically in the
types but I think I'm satisfied with the design.

```typescript
const checkDef: LiquidCheckDefinition {
  create(context) {
    context.report({
      // ...
      fix(corrector) {
        // corrector is StringCorrector
      },
      suggest: [
        {
          message: 'try this',
          fix(corrector) {
            // corrector is StringCorrector
          }
        }
      ]
    });
  }
}

const checkDef: JSONCheckDefinition {
  create(context) {
    context.report({
      // ...
      fix(corrector) {
        // corrector is JSONCorrector
      },
      suggest: [
        {
          message: 'try this',
          fix(corrector) {
            // corrector is JSONCorrector
          }
        }
      ]
    });
  }
}
```

* Add pre-wired autofix in the node package