Fix bug in autocomplete for snippets
Previously this was matching against the first period it found in the
snippet file and removing everything up until the end of the file name.

This meant that if you had a file like `foo.bar.baz.liquid` it would
only offer `foo` as the suggestion. This change allows for any number of
periods to be added to the file name and only removes the final
`.liquid`.